Issues 356 - Setup wifi connection - Removing tornado dependency - Merging WS reader with Wifi process
