+++ mycroft/configuration/mycroft.ini
@@ -111,5 +111,3 @@ ap_iface_ip = '172.24.1.1'
 ap_iface_ip_range_start = '172.24.1.10'
 ap_iface_ip_range_stop = '172.24.1.20'
 ap_iface_mac = 'bc:5f:f4:be:7d:0a'
-http_port = '8888'
-ws_port = '80'
